enhance: Abstract Execute interface for import/preimport task (#33234)
